+++ CHANGELOG.md
@@ -33,6 +33,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `digits()` gone in favor of using `usize::checked_ilog10()`
 * `rust_version` and `MSRV`: these crates aim to support the latest stable version of rust
 
+### Changed
+* [breaking-python] The `Miniconf::create` method is no longer used. Instead, an `aiomqtt::Client`
+must be passed to Miniconf
+
 ## [0.11.0](https://github.com/quartiq/miniconf/compare/v0.10.1...v0.11.0) - 2024-04-30
 
 ### Changed